Add : Releases page
- お知らせページを追加
- コンポーザブルを整理
