+++ composables/supabase.ts
@@ -1,4 +1,4 @@
-import type { SupabaseClient } from "@supabase/supabase-js";
+import type { SupabaseClient } from '@supabase/supabase-js';
 
 let client: SupabaseClient | undefined;
 
@@ -7,17 +7,4 @@ export const useSBClient = async () => {
         client = await useSupabaseClient();
     }
     return client;
-};
-
-export const useGetPopularTags = async () => {
-    const client = await useSBClient();
-
-    const { data, error } = await client.rpc("tags_order_by_count");
-
-    if (!error) {
-        return data.map((obj: { tag: string }) => obj.tag);
-    } else {
-        console.error(error);
-        return null;
-    }
 };
